drivers: add a crc7 interface and software implementation.
crc7 is used by the MMC/SD card interface and the WINC1500 Wifi
controller.  This adds a bytewise software implementation which is
reasonably fast and significantly more compact than a syndrome based
implementation.

The API matches Linux.  The test cases match an example SD card
packet,
http://reveng.sourceforge.net/crc-catalogue/1-15.htm#crc.cat-bits.7,
and give the same output as the Linux version.