netusb: wait for the transfer to complete before retrying.
In the same way as cdc_acm, set a semaphore in the transfer complete
interrupt and wait for it in the main loop before retrying.

This fixes a bug with >= 64 byte frames when logging is off.  If
logging is on, then the time to print logging is enough for the
transmit to complete.  If logging is off, then the 10 retries complete
quickly and try_write() gives up.

Signed-off-by: Michael Hope <mlhx@google.com>