drivers: console: add a QingKe Debug Module console driver
The QingKe V2A, V2C, and V4C have a Debug Module with two data
registers `data0` and `data1`. These can be used with SDI compatible
debuggers such as `minichlink` to send console data to the host.

Add a console driver that buffers up to three characters in data0,
communicates with the host, and fails gracefully if the driver is
enabled but the debugger is not connected. The buffering is
particularly important on hosts that use USB passthrough such as
Chrome OS.

Compared to the UART console, the debug console uses 916 bytes less
flash and frees up a UART.

Tested on the ch32v003evt and ch32v006evt. To test, add `&debug
{ status = "okay"; };` to the board Devicetree, set `zephyr,console =
&debug`, and enable the driver.

Signed-off-by: Michael Hope <michaelh@juju.nz>