drivers: spi: spi_bitbang: reduce the ROM usage by making MISO optional
In the same way as mipi_dbi_spi, check if any of the spi_bitbang
instances define MISO and, if not, omit the supporting code. This is
useful on boards with limited memory with write-only devices such as
many SPI LCDs.

This reduces the ROM usage by 60 B. It didn't reduce as much as I
hoped due to how the half duplex support is implemented.

Signed-off-by: Michael Hope <mlhx@google.com>