drivers: pwm: add support for the CH32V Advanced-control Timer (ADTM)
The ADTM is a superset of the GPTM and has the same core register
layout. See section 11.2.2 of the CH32V003RM for details on the
differences.

Use the same technique as the GD32 driver, i.e. use the same basic
driver with tweaks based on the features.

Add the GTPM and ADTM Devicetree and pinctrl entries for the 20x and
30x series.

Tested on the ch32v006evt and linkw by running
`samples/basic/blinky_pwm` and `tests/drivers/pwm/pwm_api` and
checking the output with a logic analyser.

Signed-off-by: Michael Hope <michaelh@juju.nz>